refactor(generate): rename `checker.ts` to `config_validator.ts` and convert warnings to errors for stricter validation
